requirements.txt
Brotli==1.0.9
click==8.0.3
cycler==0.11.0
dash==2.1.0
dash-bootstrap-components==1.0.2
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.2
Flask-Caching==1.3.3
Flask-Compress==1.10.1
fonttools==4.29.1
gunicorn==20.0.4
imageio==2.15.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
packaging==21.3
pathlib==1.0.1
pandas==1.4.0
patsy==0.5.2
Pillow==9.0.1
plotly==5.6.0
plotly-express==0.4.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
statsmodels==0.13.2
tenacity==8.0.1
threadpoolctl==3.1.0
tifffile==2022.2.9
Werkzeug==2.0.3
dash-extensions